pyproject.toml
[project]
name = "financial-assistant"
version = "0.1.0"
description = "Personal finance tracker with AI advisor"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "fastapi>=0.115",
    "uvicorn[standard]>=0.30",

    "langchain>=0.3",
    "langchain-openai>=0.2",
    "langchain-community>=0.3",
    "supabase>=2.7",
    "pydantic-settings>=2.4",
    "python-dateutil>=2.9",
    "httpx>=0.27",
    "duckduckgo-search>=6",
    "ddgs>=9.11.4",
    "PyJWT>=2.8",
]

[project.optional-dependencies]
dev = [
    "pytest>=8",
    "pytest-asyncio>=0.24",
    "httpx>=0.27",
]

[tool.pytest.ini_options]
pythonpath = ["."]
asyncio_mode = "strict"
